Revert and improve
